Consolidate Google Play Store upload functionality into android_setup.rb
- Moved Google Play Store upload implementation from separate commons file directly into android_setup.rb
- Removed the separate smf_upload_to_play_store.rb file to simplify the codebase structure
- Maintained all existing functionality including APK/AAB detection, rollout percentages, and release notes generation
